feat(phase-k): Knowledge Graph 6 tables (ADR-011 / ADR-012)
Phase K — KG schema 落地. Admin CRUD endpoints 留给后续 PR.

Backend:
- 6 new ORMs (genpano_models/kg.py):
  - KgCategory: industry → category 树状 taxonomy
  - KgBrand: 1:1 mapping to brands (ADR-011, brand_id UNIQUE)
  - KgProduct: 1:1 mapping to products
  - KgBrandRelation: COMPETES_WITH / SAME_GROUP，
    check brand_a_id < brand_b_id 防止双向重复
  - KgProductRelation: COMPETES_WITH / SUBSTITUTES / UPGRADES_TO /
    BUDGET_ALT_OF / PAIRS_WITH
  - KgRelationCandidate: LLM 推断 staging (ADR-012),
    status pending/approved/rejected/merged
- alembic 20260505_phase_k: 6 tables + 索引

Tests (7 全过):
- categories tree (parent_id 关联)
- brand_id UNIQUE constraint
- products basic CRUD
- brand_relations: brand_a_id < brand_b_id check 强制
- product_relations: (a, b, type) UNIQUE
- relation_candidates status check (pending|approved|rejected|merged)
- relation_candidates entity_kind check (brand|product)

验证:
- pytest: 101 passed (7 phase_k + 94 既有)
- ruff check + format: All checks passed
- mypy --strict: 0 issues in 108 files
- alembic upgrade head: 8 versions clean

Note: KG admin CRUD + LLM relation_extractor 留给 Phase K.3+ 后续 PR.

下一步 Phase D: Diagnostics rule engine + diagnostics 表.

+++ backend/alembic/versions/2026_05_05_0004_phase_k_kg.py
@@ -19,11 +19,11 @@ depends_on: str | Sequence[str] | None = None
 def upgrade() -> None:
     op.create_table(
         "kg_categories",
-        sa.Column("id", sa.BigInteger, primary_key=True, autoincrement=True),
+        sa.Column("id", sa.Integer, primary_key=True, autoincrement=True),
         sa.Column("industry_id", sa.Integer, nullable=True),
         sa.Column(
             "parent_id",
-            sa.BigInteger,
+            sa.Integer,
             sa.ForeignKey("kg_categories.id", ondelete="SET NULL"),
             nullable=True,
         ),
@@ -40,7 +40,7 @@ def upgrade() -> None:
 
     op.create_table(
         "kg_brands",
-        sa.Column("id", sa.BigInteger, primary_key=True, autoincrement=True),
+        sa.Column("id", sa.Integer, primary_key=True, autoincrement=True),
         sa.Column("brand_id", sa.Integer, nullable=False, unique=True),
         sa.Column("industry_id", sa.Integer, nullable=True),
         sa.Column("primary_name", sa.String(256), nullable=False),
@@ -56,10 +56,10 @@ def upgrade() -> None:
 
     op.create_table(
         "kg_products",
-        sa.Column("id", sa.BigInteger, primary_key=True, autoincrement=True),
+        sa.Column("id", sa.Integer, primary_key=True, autoincrement=True),
         sa.Column("product_id", sa.Integer, nullable=False, unique=True),
         sa.Column("brand_id", sa.Integer, nullable=False),
-        sa.Column("category_id", sa.BigInteger, nullable=True),
+        sa.Column("category_id", sa.Integer, nullable=True),
         sa.Column("primary_name", sa.String(256), nullable=False),
         sa.Column("name_zh", sa.String(256), nullable=True),
         sa.Column("name_en", sa.String(256), nullable=True),
